requirements.txt
hypercorn==0.16.0
fastapi==0.110.0
playwright==1.50
python-dotenv==1.1.1
